+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2025-02-06]
+## [0.0.1 - 2025-02-07]
 ### Added
 - First version of the project
 - Spring Application
@@ -208,3 +208,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - App Register Service
 - Coder Param (APP_ID and APP_SECRET)
 - Relate project constraints with project state constraints for DEVELOPER, PILOT and FINAL users
+- ROOT_LOG_LEVEL for ROOT and LOG_LEVEL only for de.samply packages
+- Exporter Job and Exporter Service debug messages
+- Performant HTTP Client configuration